+++ lambdas/data-migration2/src/executions.ts
@@ -21,7 +21,6 @@ const logger = new Logger({ sender: '@cumulus/data-migration/executions' });
  * @param {Knex} knex - Knex client for writing to RDS database
  * @returns {Promise<number>} - Cumulus ID for record
  * @throws {RecordAlreadyMigrated}
- *   if record was already migrated
  */
 export const migrateExecutionRecord = async (
   dynamoRecord: ExecutionRecord,
@@ -40,7 +39,6 @@ export const migrateExecutionRecord = async (
       throw error;
     }
   }
-  // Throw error if it was already migrated.
   if (existingRecord && existingRecord.updated_at >= new Date(dynamoRecord.updatedAt)) {
     throw new RecordAlreadyMigrated(`Execution arn ${dynamoRecord.arn} was already migrated, skipping`);
   }
@@ -72,6 +70,7 @@ export const migrateExecutions = async (
   const executionsTable = envUtils.getRequiredEnvVar('ExecutionsTable', env);
   const bucket = process.env.system_bucket;
   const stackName = process.env.stackName;
+  const loggingInterval = env.loggingInterval ? Number.parseInt(env.loggingInterval, 10) : 100;
 
   const searchQueue = new DynamoDbSearchQueue({
     TableName: executionsTable,
@@ -91,13 +90,16 @@ export const migrateExecutions = async (
   while (record) {
     migrationSummary.dynamoRecords += 1;
 
+    if (migrationSummary.dynamoRecords % loggingInterval === 0) {
+      logger.info(`Batch of ${loggingInterval} execution records processed, ${migrationSummary.dynamoRecords} total`);
+    }
+
     try {
       await migrateExecutionRecord(<ExecutionRecord>record, knex);
       migrationSummary.success += 1;
     } catch (error) {
       if (error instanceof RecordAlreadyMigrated) {
         migrationSummary.skipped += 1;
-        logger.info(error);
       } else {
         migrationSummary.failed += 1;
         errorMessage = `Could not create execution record in RDS for Dynamo execution arn ${record.arn}:`;