+++ lambdas/data-migration2/tests/test-granulesAndFiles.js
@@ -1,9 +1,11 @@
 const cryptoRandomString = require('crypto-random-string');
 const omit = require('lodash/omit');
+const sinon = require('sinon');
 const test = require('ava');
 
 const Granule = require('@cumulus/api/models/granules');
 const s3Utils = require('@cumulus/aws-client/S3');
+const Logger = require('@cumulus/logger');
 
 const { dynamodbDocClient } = require('@cumulus/aws-client/services');
 const { fakeFileFactory } = require('@cumulus/api/lib/testUtils');
@@ -618,3 +620,23 @@ test.serial('migrateGranulesAndFiles calls storeError when system bucket is prov
   const messageBody = JSON.parse(item.Body);
   t.truthy(messageBody.errors[0]);
 });
+
+test.serial('migrateGranulesAndFiles logs summary of migration for a specified loggingInterval', async (t) => {
+  const logSpy = sinon.spy(Logger.prototype, 'info');
+  const {
+    knex,
+    testGranule,
+  } = t.context;
+  process.env.loggingInterval = 1;
+
+  await granulesModel.create(testGranule);
+
+  t.teardown(async () => {
+    logSpy.restore();
+    await granulesModel.delete(testGranule);
+  });
+
+  await migrateGranulesAndFiles(process.env, knex);
+  t.true(logSpy.calledWith('Batch of 1 granule records processed, 1 total'));
+});
+